moved rpg.py
